Add FieldValueTrait
